Renamed and moved datasheet panel
